+++ app/views/enroll_courses/enroll_courses.json.jbuilder
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+json.course do
+  json.partial! 'courses/users', course: @course
+  json.partial! 'courses/assignments', course: @course
+end
