+++ app/views/enroll_courses/index.html.haml
@@ -1,0 +1,26 @@
+.container
+
+  %h1 Search
+
+  =form_tag users_path, method: :get, style: "width: 20%; display: inline;" do
+    = text_field_tag :email, params[:email], placeholder: 'Search for emails...'
+  =form_tag users_path, method: :get, style: "width: 20%; display: inline;" do
+    = text_field_tag :real_name, params[:real_name], placeholder: 'Search for real names...'
+
+  %hr
+
+  %h1 Users
+  %span= params[:email].present? ? "Results for: \"#{params[:email]}\"" : params[:real_name].present? ? "Results for: \"#{params[:real_name]}\"" : nil
+
+  %table.table.table--hoverable
+    %thead
+      %tr
+        %th Username
+        %th Real Name
+        %th E-mail
+    %tbody
+    - @users.each do |user|
+      %tr
+        %td= link_to user.username, user_path(user.username)
+        %td= user.real_name
+        %td= user.email
